Move non-leaflet components
